Added citations
Added references for the definition of Wigner, Kirkwood-Dirac, and SIC-POVVM representations.

Adhikary 2020 defines that:
"The state is non-classical
with respect to these observables if, even one of the
pseudo-probabilities in the scheme is negative."

That should be sufficient to support your claim that there is a focus on negative probabilities, but better references might exist. I started with Asthana 2021 and found references therein.

I properly added the citation of Hall for the entropic volume.