NonAdditiveProbability: SIC-POVM figure
Added a figure to illustrate a SIC-POVM on the Bloch ball, which is a tetrahedron inside the ball.
Alternatively, one could define a tetrahedron that fits the ball inside, which is illustrated in a separate subfigure.
There is not enough space on the figure to label all the extreme points as was done here:
https://commons.wikimedia.org/wiki/File:Regular_tetrahedron_inscribed_in_a_sphere.svg
The text would be too small.
I left the labels as comments anyway.