NonAdditiveProbability: Dumped axioms and propositions
I dumped all the axioms with their justifications and key propositions for state and fraction capacity from the book into the appendix.
I excluded the mathSection environment, but kept the axiom and justification environment.
This is just the raw material that needs to be trimmed down.

There are some inconsistencies between the book and the paper:
\size vs. \fraction
\chull vs. \hull